Round, Action State history now saved
Each attack is now stored in a double array of FightState objects
attached to a gladiator's fightItem - This allows checks of any past
activity during the fight - this allows for multi action decision making
e.g. if you missed last round you get a buff this round, and multi-round
debuffs like bleeding.  Also it will be used to store fighter position
and heading.